♻️ Refactor merge-room components to merged-contacts
- Rename merge-room/ directory to merged-contacts/
- Rename contact-merge-* components to merged-contact-*
- Update component interfaces and function names
- Fix nested ternary operators warning in SonarQube
- Improve code structure and naming consistency
